Improve BundlePreferences (Verdict Settings preference page)
Reorder the field editors so that the STEM project path and Bundle
Docker Image come first, followed by the Bundle Jar and the rest of
the non-Docker field editors.

Change init method to call setPreferencesStore correctly (the old
BundlePreferences required the use of save handlers to work around the
incorrect setPreferenceStore call).  Remove noDefaultAndApplyButton
call since the Reset to Defaults button makes clearing the non-Docker
field editors much easier when switching from binaries to Docker.

Override propertyChange (and addField) methods to iterate over the
field editors and disable all the non-Docker field editors if the user
types a Docker image name.  Now the user can enter only the STEM
project path and the docker image name; the rest of the field editors
can remain empty or be filled in if they are not disabled.

Simplify createFieldEditors to construct the field editors using fewer
arguments without needing to add save handlers and put a label between
the Docker and non-Docker field editors telling the user the remaining
fields will not be needed if Docker is used.

Add a LabelFieldEditor class to construct the label between the field
editors.

Modify CRVHandler and MBASHandler to allow empty values in the
non-Docker fields if the user has typed a Docker image name.