Display failure paths with more information
Change what the VERDICT plugin displays in the Security Failure Paths
tab.  Rename the Suggested Defense column to Suggested Defenses and
make that column contain the names of specific defenses instead of
NIST profiles.  Add a Suggested Defenses Profile column and make it
contain the NIST profiles.  Move the NIST profile tooltip to the new
column.

Update the code that reads the XML output files from Soteria++ to
handle the changes to the attributes.  Rename some methods/fields to
follow these changes or to be clearer what they represent.