Refactor into continuous, main, & release workflows
Instead of having two workflows, refactor into three workflows:

  continuous  Checks that a pull request can build the Verdict jars
  main        Builds the verdict-dev Docker image on each push to master
  release     Builds an almost complete Verdict release

We want to check a pull request and give feedback as quickly as
possible, so we build only the Verdict jars.

We want to let developers test changes in master as quickly as
possible, so we rebuild the verdict-dev image on each push to master.
Maybe later we'll copy changes to a verdict-dev update site for the
Verdict plugin too.

We don't want to build two redundant Docker images, so we build only
the verdict image when we make a release.  We'll also copy a release
to VERDICT-update-sites when we have time to implement that later.