VERDICT-270:
Cutset likelihood is represented in scientific notation for likelihoods < 1e-5 but as decimals otherwise. The notation effects MBAS output which maps likelihood levels to severity levels where doubles are not expected and default to passing "E" severity.

The likelihood print function was updated to display values in e+-dd decimal notation

Bugfix: Removes trailing ":" in attack strings of the AD fragments

Solves:
VERDICT-270